+++ consts.py
@@ -1,0 +1,12 @@
+from pygame.locals import *
+
+KEYS = {
+    1: K_s,
+    2: K_d,
+    3: K_SEMICOLON,
+    4: K_QUOTE
+}
+
+HEIGHT = 540
+WIDTH = 300
+FPS = 60
